# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "auth_login",
        "parameters": [
          {
            "name": "grimity-app-model",
            "in": "header",
            "description": "앱에서만 사용되는 속성입니다",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "grimity-app-device",
            "in": "header",
            "description": "앱에서만 사용되는 속성입니다",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "mobile",
                "tablet"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "유효성 검사 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "회원 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "로그인",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "auth_register",
        "parameters": [
          {
            "name": "grimity-app-model",
            "in": "header",
            "description": "앱에서만 사용되는 속성입니다",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "grimity-app-device",
            "in": "header",
            "description": "앱에서만 사용되는 속성입니다",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "mobile",
                "tablet"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "유효성 검사 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Register409Response"
                }
              }
            }
          }
        },
        "summary": "회원가입",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "get": {
        "operationId": "auth_refresh",
        "parameters": [
          {
            "name": "grimity-app-model",
            "in": "header",
            "description": "앱에서만 사용되는 속성입니다",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "grimity-app-device",
            "in": "header",
            "description": "앱에서만 사용되는 속성입니다",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "mobile",
                "tablet"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JwtResponse"
                }
              }
            }
          },
          "400": {
            "description": "유효성 검사 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "토큰 재발급 - refT를 담아야함",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "auth_logout",
        "parameters": [
          {
            "name": "grimity-app-model",
            "in": "header",
            "description": "앱에서만 사용되는 속성입니다",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "grimity-app-device",
            "in": "header",
            "description": "앱에서만 사용되는 속성입니다",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "mobile",
                "tablet"
              ]
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "유효성 검사 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "로그아웃 - refT를 담아야함",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/albums": {
      "post": {
        "operationId": "album_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAlbumRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdResponse"
                }
              }
            }
          },
          "400": {
            "description": "유효성 검사 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "앨범 이름 중복",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "앨범 생성",
        "tags": [
          "albums"
        ]
      }
    },
    "/albums/order": {
      "put": {
        "operationId": "album_updateOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAlbumOrderRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "유효성 검사 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "앨범 순서 변경",
        "tags": [
          "albums"
        ]
      }
    },
    "/albums/null": {
      "put": {
        "operationId": "album_removeFeeds",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveFeedsAlbumRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "유효성 검사 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "앨범에서 피드 빼기",
        "tags": [
          "albums"
        ]
      }
    },
    "/albums/{id}": {
      "put": {
        "operationId": "album_insertFeeds",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InsertFeedsRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "유효성 검사 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "앨범에 피드 넣기",
        "tags": [
          "albums"
        ]
      },
      "patch": {
        "operationId": "album_updateOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAlbumRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "유효성 검사 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "앨범 이름 중복",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "앨범 수정",
        "tags": [
          "albums"
        ]
      },
      "delete": {
        "operationId": "album_deleteOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "유효성 검사 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "앨범 삭제",
        "tags": [
          "albums"
        ]
      }
    },
...
  },
  "info": {
    "title": "grimity API",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "LoginRequest": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "enum": [
              "GOOGLE",
              "KAKAO",
              "APPLE"
            ],
            "description": "대소문자 구분 X"
          },
          "providerAccessToken": {
            "type": "string"
          },
          "deviceId": {
            "type": "string",
            "description": "디바이스 ID, 앱에서만 씁니다"
          }
        },
        "required": [
          "provider",
          "providerAccessToken"
        ]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "id": {
            "type": "string"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "id"
        ]
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "enum": [
              "GOOGLE",
              "KAKAO",
              "APPLE"
            ],
            "description": "대소문자 구분 X"
          },
          "providerAccessToken": {
            "type": "string"
          },
          "deviceId": {
            "type": "string",
            "description": "디바이스 ID, 앱에서만 씁니다"
          },
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 12
          },
          "url": {
            "type": "string",
            "minLength": 2,
            "maxLength": 20,
            "description": "id필드말고 url 필드를 써주세요"
          }
        },
        "required": [
          "provider",
          "providerAccessToken",
          "name"
        ]
      },
      "Register409Response": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "enum": [
              409
            ]
          },
          "message": {
            "type": "string",
            "enum": [
              "NAME",
              "URL"
            ]
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
      "JwtResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "accessToken",
          "refreshToken"
        ]
      },
...
      "IdResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      },
...
      "CreateAlbumRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 15,
            "description": "앨범 개수는 최대 8개"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateAlbumOrderRequest": {
        "type": "object",
        "properties": {
          "ids": {
            "description": "앨범 ID들만 배열로 담아서 주시면됩니다",
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 8,
              "minLength": 2
            }
          }
        },
        "required": [
          "ids"
        ]
      },
      "RemoveFeedsAlbumRequest": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            }
          }
        },
        "required": [
          "ids"
        ]
      },
      "InsertFeedsRequest": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            }
          }
        },
        "required": [
          "ids"
        ]
      },
      "UpdateAlbumRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 15,
            "description": "앨범 개수는 최대 8개"
          }
        },
        "required": [
          "name"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "description": "유효성 검사 실패 시 문자열 배열로 내려온다"
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      }
    }
  }
}